fix compile issues on linux

+++ libs/binder/IMemory.cpp
@@ -297,11 +297,11 @@ void BpMemoryHeap::assertReallyMapped() const
         uint32_t offset = reply.readInt32();
 
         ALOGE_IF(err, "binder=%p transaction failed fd=%d, size=%ld, err=%d (%s)",
-                asBinder().get(), parcel_fd, size, err, strerror(-err));
+                asBinder().get(), parcel_fd, (long)size, err, strerror(-err));
 
         int fd = dup( parcel_fd );
         ALOGE_IF(fd==-1, "cannot dup fd=%d, size=%ld, err=%d (%s)",
-                parcel_fd, size, err, strerror(errno));
+                parcel_fd, (long)size, err, strerror(errno));
 
         int access = PROT_READ;
         if (!(flags & READ_ONLY)) {
@@ -314,7 +314,7 @@ void BpMemoryHeap::assertReallyMapped() const
             mBase = mmap(0, size, access, MAP_SHARED, fd, offset);
             if (mBase == MAP_FAILED) {
                 ALOGE("cannot map BpMemoryHeap (binder=%p), size=%ld, fd=%d (%s)",
-                        asBinder().get(), size, fd, strerror(errno));
+                        asBinder().get(), (long)size, fd, strerror(errno));
                 close(fd);
             } else {
                 mSize = size;
